requirements.txt
google-cloud-bigquery
pandas
firebase-admin